roll game

+++ cogs/rolls.py
@@ -55,14 +55,16 @@ class rollsCog(commands.Cog, name="rolls"):
     @commands.command()
     @commands.has_role("cumdev")
     async def ct(self,ctx):
-        async with aiosqlite.connect('rolls.db') as db:
-                await db.execute("CREATE TABLE rolltable (user INT UNIQUE, alias TEXT, rolls INT, totalrolls INT, cums INT, borpas INT, goldborpaspins INT)")
-                await db.commit()
-                print("table created")
+        if str(ctx.channel.id) == '853625002779869204' or '562352225423458326':
+            async with aiosqlite.connect('rolls.db') as db:
+                    await db.execute("CREATE TABLE rolltable (user INT UNIQUE, alias TEXT, rolls INT, totalrolls INT, cums INT, borpas INT, goldborpaspins INT)")
+                    await db.commit()
+                    print("table created")
 
     @commands.command()
     @commands.has_role("cumdev")
     async def dt(self,ctx):
+        if str(ctx.channel.id) == '853625002779869204' or '562352225423458326':
             async with aiosqlite.connect('rolls.db') as db:
                 await db.execute("DROP TABLE rolltable")
                 await db.commit()
@@ -71,6 +73,7 @@ class rollsCog(commands.Cog, name="rolls"):
     @commands.command()
     @commands.has_role("cumdev")
     async def setroll(self,ctx, arg1, arg2):
+        if str(ctx.channel.id) == '853625002779869204' or '562352225423458326':
             async with aiosqlite.connect('rolls.db') as db:
                 try:
                     await db.execute("UPDATE rolltable SET rolls=? WHERE user=?",(arg1,arg2))
@@ -84,7 +87,7 @@ class rollsCog(commands.Cog, name="rolls"):
 
     @commands.command()
     async def cum(self, ctx):
-        if str(ctx.channel.id) == '262371002577715201' or '562352225423458326':
+        if str(ctx.channel.id) == '853625002779869204' or '562352225423458326':
             await ctx.message.add_reaction(emoji=loading) 
             data = await dbget()
             playerid = ctx.author.id
@@ -118,7 +121,7 @@ class rollsCog(commands.Cog, name="rolls"):
 
     @commands.command()
     async def cum10(self, ctx):
-        if str(ctx.channel.id) == '262371002577715201' or '562352225423458326':
+        if str(ctx.channel.id) == '853625002779869204' or '562352225423458326':
             await ctx.message.add_reaction(emoji=loading) 
             data = await dbget()
             playerid = ctx.author.id
@@ -155,7 +158,7 @@ class rollsCog(commands.Cog, name="rolls"):
 
     @commands.command()
     async def howmanycumsdoihaveleft(self, ctx):
-        if str(ctx.channel.id) == '262371002577715201' or '562352225423458326':
+        if str(ctx.channel.id) == '853625002779869204' or '562352225423458326':
             playerid = ctx.author.id
             data = await dbget()
             try:
@@ -164,49 +167,104 @@ class rollsCog(commands.Cog, name="rolls"):
             except:
                 await ctx.reply("You're not on the cummies list, please cum once first.")
 
+    
+    #@commands.command()
+    #async def donate(self, ctx, arg1, arg2):
+    #    data = await dbget()
+
+    #GAMES
+
+    @commands.command()
+    async def dice(self, ctx, arg):
+        if str(ctx.channel.id) == '853625002779869204' or '562352225423458326':
+            data = await dbget()
+            #try:
+            playerid = ctx.author.id
+            playername = ctx.author.name
+            match = next((item for item in data if item['user'] == playerid), 'Nothing Found')
+            if int(match['rolls']) >= int(arg):
+                #gamba
+                db = await aiosqlite.connect('rolls.db')
+                await db.execute("UPDATE rolltable SET rolls=rolls-? WHERE user=?",(arg,playerid))
+                await db.commit()
+                embed = discord.Embed(title="Roll the dice!",color=0x9062d3)
+                embed.add_field(name="Ullecbot", value="rolling...", inline=True)
+                embed.add_field(name=playername, value="rolling...", inline=True)
+                sent = await ctx.reply(embed=embed)
+
+                playerrolls=[]
+                botrolls=[]
+                for i in range(3):
+                    await asyncio.sleep(1)
+                    x = random.randint(1,6)
+                    y = random.randint(1,6)
+                    botrolls.append(x)
+                    playerrolls.append(y)
+                    displaybot = ', '.join(map(str,botrolls))
+                    displayplayer = ', '.join(map(str,playerrolls))
+                    embed.set_field_at(0, name="Ullecbot", value=displaybot, inline=True)
+                    embed.set_field_at(1, name=playername, value=displayplayer, inline=True)
+                    await sent.edit(embed=embed)
+                botsum = sum(botrolls)
+                playersum = sum(playerrolls)
+                if botsum > playersum:
+                    embed.add_field(name='You lose!!', value="You lost "+ arg +" rolls", inline=False)
+                if playersum > botsum:
+                    embed.add_field(name='You win!!', value="You win "+ arg +" rolls", inline=False)
+                    arg = int(arg) + int(arg)
+                    print(arg)
+                    await db.execute("UPDATE rolltable SET rolls=rolls+? WHERE user=?",(arg,playerid))
+                if botsum == playersum:
+                    embed.add_field(name='You draw!!', value="Your rolls have been returned", inline=False)
+                    await db.execute("UPDATE rolltable SET rolls=rolls+? WHERE user=?",(arg,playerid))
+                await db.commit()
+                await sent.edit(embed=embed)
+            else:
+                await ctx.reply("You dont have enough rolls for this.")
+        #except:
+        #    print("something went wrong")
+
 
     #STATS
 
     @commands.command()
     async def cumsavers(self, ctx):
-        db = await aiosqlite.connect('rolls.db')
-        cursor = await db.execute('SELECT alias, rolls FROM rolltable ORDER BY rolls DESC LIMIT 5')
-        rows = await cursor.fetchall()
-        ctopid = await db.execute("SELECT user FROM rolltable ORDER BY rolls DESC LIMIT 1")
-        topid = await ctopid.fetchone()
-        top5 = [{'user': a,'rolls': b,} for a,b in rows]
-        display = [("%s: %s cums saved"%(item['user'],item['rolls'])) for item in top5]
-        sep = '\n'
-        embed = discord.Embed(title="Top Cum Savers!",color=0x9062d3)
-        top = self.bot.get_user(topid[0])
-        embed.set_thumbnail(url=top.avatar_url)
-        embed.add_field(name="​", value=sep.join(display), inline=False)
-        await ctx.send(embed=embed)
+        if str(ctx.channel.id) == '853625002779869204' or '562352225423458326':
+            db = await aiosqlite.connect('rolls.db')
+            cursor = await db.execute('SELECT alias, rolls FROM rolltable ORDER BY rolls DESC LIMIT 5')
+            rows = await cursor.fetchall()
+            ctopid = await db.execute("SELECT user FROM rolltable ORDER BY rolls DESC LIMIT 1")
+            topid = await ctopid.fetchone()
+            top5 = [{'user': a,'rolls': b,} for a,b in rows]
+            display = [("%s: %s cums saved"%(item['user'],item['rolls'])) for item in top5]
+            sep = '\n'
+            embed = discord.Embed(title="Top Cum Savers!",color=0x9062d3)
+            top = self.bot.get_user(topid[0])
+            embed.set_thumbnail(url=top.avatar_url)
+            embed.add_field(name="​", value=sep.join(display), inline=False)
+            await ctx.send(embed=embed)
 
     @commands.command()
     async def cumstats(self, ctx):
-        db = await aiosqlite.connect('rolls.db')
-        playerid = ctx.author.id
-        cursor = await db.execute('SELECT rolls, totalrolls, cums, borpas, goldborpaspins FROM rolltable WHERE user=?',[playerid])
-        stats = await cursor.fetchall()
-        print(stats)
-        stats = [{'rolls': b, 'total rolls': c, 'cums': d, 'borpaspins': e, 'goldborpaspins': f,} for b,c,d,e,f in stats]
-        print(stats)
-        print(stats[0])
-        listedstats =[]
+        if str(ctx.channel.id) == '853625002779869204' or '562352225423458326':
+            db = await aiosqlite.connect('rolls.db')
+            playerid = ctx.author.id
+            cursor = await db.execute('SELECT rolls, totalrolls, cums, borpas, goldborpaspins FROM rolltable WHERE user=?',[playerid])
+            stats = await cursor.fetchall()
+            stats = [{'rolls saved': b, 'total rolls': c, 'cums': d, 'borpaspins': e, 'goldborpaspins': f,} for b,c,d,e,f in stats]
+            listedstats =[]
+            for i,o in stats[0].items():
+                line = i + ': ' + str(o)
+                listedstats.append(line)
+            #room for more stats later % etc
 
-        for i,o in stats[0].items():
-            line = i + ': ' + str(o)
-            listedstats.append(line)
-        
-
-        sep = '\n'
-        final = sep.join(listedstats)
-        titlestring = ctx.author.name + "'s stats"
-        embed = discord.Embed(title=titlestring,color=0x9062d3)
-        embed.set_thumbnail(url=ctx.author.avatar_url)
-        embed.add_field(name="​", value=final, inline=False)
-        await ctx.send(embed=embed)
+            sep = '\n'
+            final = sep.join(listedstats)
+            titlestring = ctx.author.name + "'s stats"
+            embed = discord.Embed(title=titlestring,color=0x9062d3)
+            embed.set_thumbnail(url=ctx.author.avatar_url)
+            embed.add_field(name="​", value=final, inline=False)
+            await ctx.send(embed=embed)
     
     
     #LISTENERS
